Move functions from Map Component to mapUtils file
